requirements.txt
pytest~=8.3.2
requests~=2.32.3
allure-python-commons~=2.13.5